refactor: move app state related tests to app_state test mod
